pyproject.toml
[tool.poetry]
name = "dane-audio-extraction-worker"
version = "0.1.0"
description = "Extract audio fragments from AV files given a list of time codes"
authors = ["Willem Melder <wmelder@beeldengeluid.nl>", "Jaap Blom <jblom@beeldengeluid.nl>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
pika = "^1.3.1"
yacs = "^0.1.8"
dane = "^0.3.7"
ffmpeg-python = "^0.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
mockito = "^1.3.3"
flake8 = "^5.0.4"
pytest-cov = "^4.0.0"
black = "^22.10.0"
mypy = "^0.982"
types-requests = "^2.28.11.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
  "tests/unit",
]
addopts = [
  "--cov",
]

[tool.coverage.run]
omit = [
  "tests/*"
]

[tool.black]
line-length = 88
target_version = ['py310']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | build
    | dist
  )/
)
'''

[tool.mypy]
python_version = "3.10"

[[tool.mypy.overrides]]
module = [
  'dane.*',
  'mockito',
  'pika',
  'pika.exceptions',
  'yaml',
  'yacs.*',
]
ignore_missing_imports = true
